Don't `describe` all tables
Running `describe <table>` in a loop is a very expensive
operation, especially for a database with a lot of tables.
Instead, we now query the information schema and build the
relevant descriptions for all the tables in one query.

Furthermore, the `table_cache` was not being populated and
the table completions were being populated from the `keywords`'s
help documentation and not from the `table_cache`. That is
also now fixed.

Also adds a few type corrections and changes to make the
codebase better.