feat: Allow connection switch and query execute
Add ability to switch connections on the fly. This will
also update all the help strings without having to reconnect
the server.

Query execution worked by selecting the query and running
execute on it. It now works simply by running the code action
while having the cursor in the query.